src/executor/requirements.txt
pyautogui
pillow
pyperclip
autopep8
# その他、必要に応じて追加

